ADD repositories for toy images & Product + relational configuration between the two

+++ src/main/java/com/example/toystorey/repositories/ProductRepository.java
@@ -1,16 +1,12 @@
 package com.example.toystorey.repositories;
 
 import com.example.toystorey.model.Product;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
+public interface ProductRepository extends PagingAndSortingRepository<Product, Long> {
     List<Product> findAllByName(String name, Pageable pageable);
-
     List<Product> findAllByPrice(double price, Pageable pageable);
-
 }
